unsignedBigInteger en vez de integer

+++ database/migrations/2025_02_04_000001_add_familia_id_to_ciclos_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('ciclos', function (Blueprint $table) {
-            $table->integer('familia_id')
+            $table->unsignedBigInteger('familia_id')
                 ->nullable()
                 ->after('codFamilia');
             $table->foreign('familia_id')
